+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Featherlight – Changelog
 
 Master
 -----------------------------------
-
+### Features
+- New setting: loading (default is '') is shown initially while content loads. The lightbox also has a class '.featherlight-loading' while content is loading.
 
 1.1.0 - 2015-02-17
 -----------------------------------